Extract container & event attributes to constants
Since these are shared between a couple of the modules, extract
these here for reuse.
